requirements.txt
Flask>=3.0.0
gunicorn>=21.2.0
authlib>=1.3.0
google-auth>=2.0.0
google-api-python-client>=2.0.0
